Added repository interface for separation logic from implementation.
