+++ app/src/main/java/com/tromian/game/afproject/presentation/viewmodels/MoviesViewModelProviderFactory.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModelProvider
 import com.tromian.game.afproject.domain.repository.MoviesRepository
 
 class MoviesViewModelProviderFactory(
-        val app : Application,
-        private val repository : MoviesRepository
+    val app : Application,
+    private val repository : MoviesRepository
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return MoviesViewModel(app, repository) as T
